added more OOP
