+++ src/main/java/bhavs.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 class bhavs {
     private static final String FILE_PATH = "./data/duke.txt"; // File path for saving tasks (not sure how to give this)
-    private List<Task> taskList; // List of tasks
+    private TaskList taskList; // List of tasks
 
     public static void main(String[] args) {
         bhavs chatBot = new bhavs();
@@ -12,12 +12,12 @@ class bhavs {
     }
 
     public bhavs() {
-        this.taskList = new ArrayList<>();
+        this.taskList = new TaskList();
     }
 
     public void run() {
         UI ui = new UI(FILE_PATH, taskList);
-        ui.loadTasksFromFile();
+        ui.loadTasks();
         ui.printWelcomeMessage();
         ui.personalWelcomeToGuest();
         ui.processComands();
